requirements.txt
Flask~=2.0.1
matplotlib~=3.4.2
torch~=1.8.1
torchvision~=0.9.1
numpy~=1.20.3
Werkzeug~=2.0.1
Pillow~=8.2.0